pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "patron"
version = "2.0.1"
description = "Telegram Agentic AI bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain~=1.2.12",
    "langchain-google-genai",
    "psycopg[binary]~=3.3.3",
    "langgraph-checkpoint-mongodb~=0.3.1",
    "python-telegram-bot[job-queue]~=22.5",
    "python-dotenv~=1.2",
    "pytest",
    "pytest-asyncio",
    "flake8",
    "injector~=0.24",
    "requests",
    "qdrant-client~=1.17.1",
    "pymongo~=4.15.5",
    "mongomock~=4.3",
    "croniter~=6.2",
]

[tool.setuptools.packages.find]
where = ["src"]
